Add match management and attendance features for staff users
